% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catalog_table}
\alias{catalog_table}
\title{Surface Elevation Catalog}
\format{
An object of class \code{data.frame} with 99 rows and 12 columns.
}
\source{
OWP Spatial
}
\usage{
catalog_table
}
\description{
Table of CONUS wide elevation sources
}
\keyword{datasets}
